Have clases return Error trait rather thatn string

+++ src/models/training_result.rs
@@ -1,4 +1,4 @@
 pub struct TrainingResult {
     pub id: String,
-    pub elapsed: i64,
+    pub elapsed: u128,
 }
